backend/requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6

# Graph Visualization
networkx>=2.8.0  # For graph operations
pydot>=1.4.2     # For graph visualization
graphviz>=0.20.0 # For layout algorithms

# Validation
email-validator>=2.0.0  # For email validation

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0
pymdown-extensions>=10.0.0
mkdocstrings>=0.22.0
mkdocs-minify-plugin>=0.7.0

# AI Integration
openai>=1.3.0

# Authentication
python-jose[cryptography]>=3.3.0  # For JWT tokens
passlib[bcrypt]>=1.7.4           # For password hashing

# Testing
pytest>=7.4.0
httpx>=0.24.1  # Required for FastAPI TestClient
